This commit adopts https://github.com/klauspost/compress. It's faster and much less allocation happy:

```
name                     old time/op    new time/op    delta
LoadElfFile/cockroach-4     1.13s ± 0%     0.86s ± 3%  -23.87%  (p=0.000 n=15+20)
LoadElfFile/agent-4         329ms ± 1%     258ms ± 1%  -21.76%  (p=0.000 n=18+17)

name                     old speed      new speed      delta
LoadElfFile/cockroach-4   209MB/s ± 0%   275MB/s ± 3%  +31.40%  (p=0.000 n=15+20)
LoadElfFile/agent-4       218MB/s ± 1%   278MB/s ± 1%  +27.88%  (p=0.000 n=19+17)

name                     old alloc/op   new alloc/op   delta
LoadElfFile/cockroach-4     249MB ± 0%     240MB ± 0%   -3.81%  (p=0.000 n=20+19)
LoadElfFile/agent-4        75.3MB ± 0%    72.4MB ± 0%   -3.83%  (p=0.000 n=17+16)

name                     old allocs/op  new allocs/op  delta
LoadElfFile/cockroach-4      110k ± 0%       14k ± 0%  -87.56%  (p=0.000 n=20+19)
LoadElfFile/agent-4         32.1k ± 0%      3.2k ± 0%  -89.92%  (p=0.000 n=20+18)
```